+++ tools/webUI/ui/__init__.py
@@ -1,0 +1,7 @@
+from tools.webUI.ui.msst import msst
+from tools.webUI.ui.vr import vr
+from tools.webUI.ui.preset import preset
+from tools.webUI.ui.tools import tools
+from tools.webUI.ui.models import models
+from tools.webUI.ui.train import train
+from tools.webUI.ui.settings import settings
